Refactor calling of extensions to use default functions
Expands on functions to perform common actions for extensions so the extensions
don't have to implement everything themselves.

If the 'help' output of an extension includes '  <command>', the extension's
command will be used rather than the internal function.

Gets variables from the extension so it knows how to perform some of the
functions.